Improve Http API
